+++ contracts/libs/Types.sol
@@ -2,8 +2,6 @@
 
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/utils/Counters.sol";
-
 library Types {
     enum GemstoneType {
         Amethyst, //Purple
@@ -16,13 +14,13 @@ library Types {
 
     struct Gemstone {
         GemstoneType gemstone;
-        uint8 supply; //50
-        uint8 lastMintedId; //1-50, INIT IS 0
+        uint256 supply; //50
+        uint256 lastMintedId; //1-50, INIT IS 0
     }
 
     struct PurchaseInfo {
         GemstoneType gemstoneType;
-        uint8 gemId;
+        uint256 gemId;
         uint256 purchasedDate; // Set with block.timestamp
         uint256 validityDate; // Set with block.timestamp + VALIDITY_PERIOD
         bool redeemed;
